Set method in controller and wrote haversine model to calc distance

+++ src/main/java/com/bartolomeo/SecondVariant/model/Geo.java
@@ -1,23 +1,23 @@
 package com.bartolomeo.SecondVariant.model;
 
 public class Geo {
-    private float lat;
+    private double lat;
 
-    private float lng;
+    private double lng;
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLng(float lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
 }
